feat(ui): split graph page server shell
